+++ src/lib/dhcp/option_custom.cc
@@ -504,7 +504,7 @@ OptionCustom::unpack(OptionBufferConstIter begin,
 uint16_t
 OptionCustom::len() {
     // The length of the option is a sum of option header ...
-    int length = getHeaderLen();
+    size_t length = getHeaderLen();
 
     // ... lengths of all buffers that hold option data ...
     for (std::vector<OptionBuffer>::const_iterator buf = buffers_.begin();
