+++ src/lib/hooks/library_manager.cc
@@ -1,4 +1,4 @@
-// Copyright (C) 2013,2015  Internet Systems Consortium, Inc. ("ISC")
+// Copyright (C) 2013, 2015  Internet Systems Consortium, Inc. ("ISC")
 //
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -150,7 +150,7 @@ LibraryManager::registerStandardCallouts() {
 
     // Iterate through the list of known hooks
     vector<string> hook_names = ServerHooks::getServerHooks().getHookNames();
-    for (int i = 0; i < hook_names.size(); ++i) {
+    for (size_t i = 0; i < hook_names.size(); ++i) {
 
         // Look up the symbol
         void* dlsym_ptr = dlsym(dl_handle_, hook_names[i].c_str());
@@ -334,7 +334,7 @@ LibraryManager::unloadLibrary() {
         // library on all hooks.
         vector<string> hooks = ServerHooks::getServerHooks().getHookNames();
         manager_->setLibraryIndex(index_);
-        for (int i = 0; i < hooks.size(); ++i) {
+        for (size_t i = 0; i < hooks.size(); ++i) {
             bool removed = manager_->deregisterAllCallouts(hooks[i]);
             if (removed) {
                 LOG_DEBUG(hooks_logger, HOOKS_DBG_CALLS, HOOKS_CALLOUTS_REMOVED)
